+++ client/src/serveurs/Script.cpp
@@ -11,7 +11,7 @@ namespace BTTP
                 const std::vector<std::string> elements = Protocole::Messages::decouper(serialisation, BTTP_MESSAGE_ARGS_SEP);
                 if (elements.size() != 2) throw Protocole::Erreur::Messages::Deconstruction::Incomplet(serialisation, elements.size(), 2);
                 this->_nom = elements[0];
-                this->_id = (uint64_t)std::atol(elements[1].c_str());
+                this->_id = (uint32_t)std::atol(elements[1].c_str());
             }
         }
     }
